SCRUM193 Refactor Explore Page to have multiple components.
